requirements.txt
streamlit
pandas
numpy
scikit-learn
altair
pyyaml
joblib
plotly
